+++ packages/doc/src/stories/widgets/Pool.stories.tsx
@@ -25,6 +25,75 @@ export const Primary = (props: any) => {
       {...config}
       tokenList={[
         {
+          name: 'TEST1',
+          address: '0x4E59c68F7216Fe96f27f47D3fA1d8306a00aB380',
+          symbol: 'test1',
+          decimals: 18,
+          slippage: null,
+          chainId: 2201,
+          logoImg:
+            'https://images.dodoex.io/Ie7OODKd0GDZlINrR3drPUEk4js0wnlL2mtt5U_bcsk/rs:fit:160:160:0/g:no/aHR0cHM6Ly9zdG9yYWdlLmdvb2dsZWFwaXMuY29tL2RvZG8tbWVkaWEtc3RhZ2luZy9kZXYvMjIwMS8weDRlNTljNjhmNzIxNmZlOTZmMjdmNDdkM2ZhMWQ4MzA2YTAwYWIzODA.webp',
+          tokenlists: [
+            {
+              name: 'All',
+              status: 'launched',
+              __typename: 'Erc20TokenListV2',
+            },
+          ],
+          domains: [],
+          funcLabels: [],
+          attributeLabels: [],
+          __typename: 'Erc20Erc20V2List',
+        },
+        {
+          name: 'TEST2',
+          address: '0x7af9428fC096e469Fa2583B9164f4f87Ba0F75A1',
+          symbol: 'test2',
+          decimals: 18,
+          slippage: null,
+          chainId: 2201,
+          logoImg:
+            'https://images.dodoex.io/4gyRfX982GJg5T_wBJhWpS50VHfjs299hY9Q5Pf5yzI/rs:fit:160:160:0/g:no/aHR0cHM6Ly9zdG9yYWdlLmdvb2dsZWFwaXMuY29tL2RvZG8tbWVkaWEtc3RhZ2luZy9kZXYvMjIwMS8weDdhZjk0MjhmYzA5NmU0NjlmYTI1ODNiOTE2NGY0Zjg3YmEwZjc1YTE.webp',
+          tokenlists: [
+            {
+              name: 'All',
+              status: 'launched',
+              __typename: 'Erc20TokenListV2',
+            },
+          ],
+          domains: [],
+          funcLabels: [],
+          attributeLabels: [],
+          __typename: 'Erc20Erc20V2List',
+        },
+        {
+          name: 'gUSDT',
+          address: '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE',
+          symbol: 'gUSDT',
+          decimals: 18,
+          slippage: null,
+          chainId: 2201,
+          logoImg:
+            'https://images.dodoex.io/53IvjioP70fgSHG71f1pVIM87xf9oZuQixZ8O-oXwf8/rs:fit:160:160:0/g:no/aHR0cHM6Ly9zdG9yYWdlLmdvb2dsZWFwaXMuY29tL2RvZG8tbWVkaWEtc3RhZ2luZy9kZXYvMjIwMS8weGVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWVlZWU.webp',
+          tokenlists: [
+            {
+              name: 'All',
+              status: 'launched',
+              __typename: 'Erc20TokenListV2',
+            },
+          ],
+          domains: [],
+          funcLabels: [],
+          attributeLabels: [
+            {
+              key: 'non-standard-erc20',
+              __typename: 'Erc20AttributeLabelV2',
+            },
+          ],
+          __typename: 'Erc20Erc20V2List',
+        },
+
+        {
           address: '0x7D381e6a9c23A0E6969658f6B8Eba57A4Dbf93a0',
           symbol: 'USDT',
           name: 'USDT',
@@ -82,7 +151,7 @@ Primary.args = {
   routerPage: undefined,
   supportAMMV2: true,
   supportAMMV3: true,
-  onlyChainId: 688689,
+  onlyChainId: 2201,
   showSubmissionSubmittedDialog: true,
   // onlyChainId: 1,
 };